requirements.txt
accelerate==0.23.0
aiohttp==3.8.4
angle_emb==0.3.3
anthropic==0.16.0
deepspeed==0.9.2
fastapi==0.109.2
FlagEmbedding==1.2.3
flash_attn==2.2.5
httpx==0.26.0
huggingface_hub==0.20.3
InstructorEmbedding==1.0.1
jsonlines==4.0.0
mteb==1.1.1
numpy==1.18.0
openai==1.12.0
opencc==1.1.7
pandas==1.3.5
peft==0.5.0
pytrec_eval==0.5
rank_bm25==0.2.2
ray==2.9.2
Requests==2.31.0
safetensors==0.3.3
scikit_learn==0.24.2
setuptools==59.5.0
shortuuid==1.0.11
torch_npu==2.1.0
tqdm==4.64.0
transformers==4.37.1
vertexai==0.0.1
vllm==0.3.0
voyageai==0.1.7
xformers==0.0.19
